requirements.txt
backports-abc==0.5
certifi==2017.4.17
docutils==0.13.1
inflection==0.3.1
lockfile==0.12.2
luigi==2.6.1
numpy==1.12.1
pandas==0.19.2
pymongo==3.4.0
python-daemon==2.1.2
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
singledispatch==3.4.0.3
six==1.10.0
statsnba-playbyplay==0.2.0
tornado==4.5.1
wheel==0.24.0
